pyproject.toml
[project]
name = "voice-factory-backend"
version = "0.1.0"
description = "Voice -> STT -> Text2SQL -> Answer -> TTS pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "duckdb>=1.0.0",
  "pandas>=2.2.0",
  "openai>=1.0.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "gradium>=0.0.1",
]
